Some refactoring

+++ database/migrations/2021_01_27_221933_create_blog_categories_table.php
@@ -14,15 +14,15 @@ class CreateBlogCategoriesTable extends Migration
     public function up()
     {
         Schema::create('blog_categories', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('parent_id')->unsigned()->default(1);
+            $table->bigIncrements('id');
+            $table->bigInteger('parent_id')->unsigned()->default(1);
 
             $table->string('slug')->unique();
-	    $table->string('title');
-	    $table->text('description')->nullable();
-	    
+            $table->string('title');
+            $table->text('description')->nullable();
+
             $table->timestamps();
-	    $table->softDeletes();
+            $table->softDeletes();
         });
     }
